Fix and upgrade saving html source in spider and selenium abstract classes.
CoronaSpider.py - In save_source_page() when calling write_file(), the parameters order was wrong, causing an unhandled exception.
In addition, the 'return' at the end of the except part, caused no following spider on the next run of the thread, since no request was created by yield.

coronaSeleniumScraper.py - When waiting for an element, if not found an exception is thrown and the browser object is not relevant and trying to get it source caused another exception. Therefore, create another browser and save the source after waiting some time using sleep().
In addition the max time to wait should be shorter and in save_source_page() when calling write_file(), the parameters order was wrong, causing an unhandled exception