Add check that inserted value to DB is larger than max and last value from same source.
coronaDB.py - Add methods for getting last and max value of a specific source from DB.
When inserting a value to the DB, check if it is higher than last value, if not - stop and throw an exception.
In the future maybe use max check also, though last should be enough.

coronaDataFetcher.py - Minor change in DB object creation.